refactor: migrar para TypeScript, isolar dados e melhorar fluidez (UX)
Reestruturação completa para escalabilidade e fluidez:

Arquitetura
- Migração completa para TypeScript (tsconfig, ESLint typescript-eslint,
  vite.config.ts, alias @/). Typecheck, lint e build limpos.
- Dados isolados da UI em src/data/* (projects, certificates, timeline,
  navigation, achievements, socials) com tipos em src/types.
- Componentes reorganizados: páginas em src/pages, componentes por domínio
  em src/components (ui, layout, projects, certificates) e helpers em src/lib.
- Remoção de código morto: CertificateCard/Modal e OptimizedImage órfãos,
  hooks sem uso, registro de Service Worker inexistente e dep prop-types.

UX / fluidez
- Transição suave entre rotas (AnimatePresence).
- Modal reutilizável único: fecha com ESC, trava scroll do fundo, animação
  de saída, portal e acessibilidade (role=dialog, foco).
- Scroll-reveal com stagger nos cards (re-anima ao filtrar/buscar).
- Foto de perfil otimizada de 26MB para ~556KB (3024x4032 -> 600px).
