+++ index.html
@@ -170,6 +170,6 @@
                 </ul>
             </footer>
         </div>
-        <script type="module" src="/src/main.jsx"></script>
+        <script type="module" src="/src/main.tsx"></script>
     </body>
 </html>
