+++ package.json
@@ -6,9 +6,10 @@
     "type": "module",
     "scripts": {
         "dev": "vite --host",
-        "build": "vite build",
-        "build:vercel": "vite build",
-        "build:gh-pages": "VITE_BASE_PATH=/luis.hsa vite build",
+        "build": "tsc --noEmit && vite build",
+        "build:vercel": "tsc --noEmit && vite build",
+        "build:gh-pages": "tsc --noEmit && VITE_BASE_PATH=/luis.hsa vite build",
+        "typecheck": "tsc --noEmit",
         "lint": "eslint .",
         "preview": "vite preview",
         "preview:vercel": "npm run build:vercel && vite preview",
@@ -19,7 +20,6 @@
         "framer-motion": "^11.12.0",
         "geist": "^1.5.1",
         "lucide-react": "^0.462.0",
-        "prop-types": "^15.8.1",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
         "react-icons": "^5.3.0",
@@ -27,6 +27,7 @@
     },
     "devDependencies": {
         "@eslint/js": "^9.15.0",
+        "@types/node": "^25.9.2",
         "@types/react": "^18.3.12",
         "@types/react-dom": "^18.3.1",
         "@vitejs/plugin-react": "^4.3.4",
@@ -39,6 +40,8 @@
         "globals": "^15.12.0",
         "postcss": "^8.4.49",
         "tailwindcss": "^3.4.15",
+        "typescript": "^6.0.3",
+        "typescript-eslint": "^8.60.1",
         "vite": "^6.0.1"
     }
 }
