+++ src/data/certificates.ts
@@ -1,21 +1,7 @@
-import { useState } from 'react';
-import { motion } from 'framer-motion';
-import {
-    Search,
-    Award,
-    Calendar,
-    ExternalLink,
-    X,
-    BookOpen,
-    Users,
-    Code2,
-    GraduationCap,
-    Clock,
-    FileText
-} from 'lucide-react';
-import JECETImage from '../assets/academicAchievements/JECET.jpg';
+import type { Certificate } from '@/types';
+import JECETImage from '@/assets/academicAchievements/JECET.jpg';
 
-const certificatesData = [
+export const certificates: Certificate[] = [
     {
         id: 1,
         title: 'UC Berkeley AI Hackathon',
@@ -322,309 +308,3 @@ const certificatesData = [
         image: 'https://styles.redditmedia.com/t5_2scpm/styles/communityIcon_72kalishkfl81.png'
     }
 ];
-
-const CertificateCard = ({ certificate, onClick }) => {
-    const getIcon = type => {
-        switch (type) {
-            case 'courses':
-                return <BookOpen size={24} className="text-gray-400" />;
-            case 'honors':
-                return <Award size={24} className="text-gray-400" />;
-            case 'extracurricular':
-                return <Users size={24} className="text-gray-400" />;
-            case 'events':
-                return <Calendar size={24} className="text-gray-400" />;
-            case 'hackathons':
-                return <Code2 size={24} className="text-gray-400" />;
-            default:
-                return <GraduationCap size={24} className="text-gray-400" />;
-        }
-    };
-
-    return (
-        <motion.div
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            whileHover={{ y: -5 }}
-            onClick={() => onClick(certificate)}
-            className="bg-dark-secondary/30 border border-gray-800 rounded-2xl p-6 cursor-pointer hover:border-gray-600 transition-all duration-300 group h-full flex flex-col"
-        >
-            <div className="flex items-start justify-between mb-6">
-                <div className="w-12 h-12 rounded-xl bg-white/5 p-2 border border-gray-800 group-hover:border-gray-600 transition-colors flex items-center justify-center overflow-hidden">
-                    {certificate.image ? (
-                        <img
-                            src={certificate.image}
-                            alt={certificate.institution}
-                            className={`w-full h-full object-cover opacity-80 group-hover:opacity-100 transition-opacity ${
-                                certificate.imagePosition || ''
-                            }`}
-                        />
-                    ) : (
-                        getIcon(certificate.type)
-                    )}
-                </div>
-                <span className="text-xs font-light text-gray-500 border border-gray-800 rounded-full px-3 py-1">
-                    {certificate.date}
-                </span>
-            </div>
-
-            <h3 className="text-lg font-medium text-white mb-1 group-hover:text-gray-200 transition-colors line-clamp-2">
-                {certificate.title}
-            </h3>
-            <p className="text-sm text-gray-400 font-light mb-4 line-clamp-2">
-                {certificate.institution}
-            </p>
-
-            <div className="mt-auto pt-4 border-t border-gray-800/50 flex justify-between items-center">
-                <span className="text-xs text-gray-500 font-light capitalize flex items-center gap-1">
-                    {certificate.type === 'courses' && <BookOpen size={12} />}
-                    {certificate.type === 'extracurricular' && (
-                        <Users size={12} />
-                    )}
-                    {certificate.type === 'honors' && <Award size={12} />}
-                    {certificate.type === 'events' && <Calendar size={12} />}
-                    {certificate.type === 'hackathons' && <Code2 size={12} />}
-                    {certificate.type}
-                </span>
-                {certificate.hours && (
-                    <span className="text-xs text-gray-500 font-light flex items-center gap-1">
-                        <Clock size={12} /> {certificate.hours}h
-                    </span>
-                )}
-            </div>
-        </motion.div>
-    );
-};
-
-const CertificateModal = ({ certificate, onClose }) => {
-    const getIcon = type => {
-        switch (type) {
-            case 'courses':
-                return <BookOpen size={32} className="text-gray-400" />;
-            case 'honors':
-                return <Award size={32} className="text-gray-400" />;
-            case 'extracurricular':
-                return <Users size={32} className="text-gray-400" />;
-            case 'events':
-                return <Calendar size={32} className="text-gray-400" />;
-            case 'hackathons':
-                return <Code2 size={32} className="text-gray-400" />;
-            default:
-                return <GraduationCap size={32} className="text-gray-400" />;
-        }
-    };
-
-    return (
-        <div
-            className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center z-50 p-4"
-            onClick={onClose}
-        >
-            <motion.div
-                initial={{ scale: 0.95, opacity: 0 }}
-                animate={{ scale: 1, opacity: 1 }}
-                className="bg-dark-primary border border-gray-800 rounded-2xl max-w-2xl w-full max-h-[90vh] overflow-y-auto shadow-2xl"
-                onClick={e => e.stopPropagation()}
-            >
-                <div className="p-8">
-                    <div className="flex justify-between items-start mb-6">
-                        <div className="flex gap-4 w-full">
-                            <div className="w-16 h-16 rounded-xl bg-white/5 p-3 border border-gray-800 flex items-center justify-center overflow-hidden shrink-0">
-                                {certificate.image ? (
-                                    <img
-                                        src={certificate.image}
-                                        alt={certificate.institution}
-                                        className={`w-full h-full object-cover ${
-                                            certificate.imagePosition || ''
-                                        }`}
-                                    />
-                                ) : (
-                                    getIcon(certificate.type)
-                                )}
-                            </div>
-                            <div className="flex-1">
-                                <h2 className="text-2xl font-light text-white mb-1">
-                                    {certificate.title}
-                                </h2>
-                                <p className="text-gray-400 font-light">
-                                    {certificate.institution} •{' '}
-                                    {certificate.date}
-                                </p>
-                            </div>
-                        </div>
-                        <button
-                            onClick={onClose}
-                            className="text-gray-500 hover:text-white transition-colors"
-                        >
-                            <X size={24} />
-                        </button>
-                    </div>
-
-                    <div className="space-y-6">
-                        <div>
-                            <h3 className="text-sm font-medium text-gray-500 uppercase tracking-wider mb-3">
-                                Details
-                            </h3>
-                            <div className="grid grid-cols-2 gap-4 mb-4">
-                                <div className="bg-dark-secondary/30 border border-gray-800 rounded-xl p-3">
-                                    <span className="text-xs text-gray-500 block mb-1">
-                                        Type
-                                    </span>
-                                    <span className="text-white font-light capitalize">
-                                        {certificate.type}
-                                    </span>
-                                </div>
-                                {certificate.hours && (
-                                    <div className="bg-dark-secondary/30 border border-gray-800 rounded-xl p-3">
-                                        <span className="text-xs text-gray-500 block mb-1">
-                                            Duration
-                                        </span>
-                                        <span className="text-white font-light">
-                                            {certificate.hours} hours
-                                        </span>
-                                    </div>
-                                )}
-                            </div>
-
-                            {certificate.description && (
-                                <div className="mt-4">
-                                    <h3 className="text-sm font-medium text-gray-500 uppercase tracking-wider mb-3">
-                                        Description
-                                    </h3>
-                                    <p className="text-gray-300 font-light leading-relaxed whitespace-pre-line">
-                                        {certificate.description}
-                                    </p>
-                                </div>
-                            )}
-                        </div>
-
-                        {(certificate.newsArticle || certificate.pdfPath) && (
-                            <div className="pt-6 border-t border-gray-800 flex flex-col gap-3">
-                                {certificate.newsArticle && (
-                                    <a
-                                        href={certificate.newsArticle}
-                                        target="_blank"
-                                        rel="noopener noreferrer"
-                                        className="inline-flex items-center gap-2 px-6 py-3 bg-dark-secondary/30 border border-gray-800 text-white rounded-full font-medium hover:bg-dark-secondary/50 transition-colors w-full justify-center"
-                                    >
-                                        <ExternalLink size={18} />
-                                        Read News Article
-                                    </a>
-                                )}
-                                {certificate.pdfPath && (
-                                    <a
-                                        href={certificate.pdfPath}
-                                        target="_blank"
-                                        rel="noopener noreferrer"
-                                        className="inline-flex items-center gap-2 px-6 py-3 bg-white text-black rounded-full font-medium hover:bg-gray-200 transition-colors w-full justify-center"
-                                    >
-                                        <FileText size={18} />
-                                        View Certificate PDF
-                                    </a>
-                                )}
-                            </div>
-                        )}
-                    </div>
-                </div>
-            </motion.div>
-        </div>
-    );
-};
-
-const Certificates = () => {
-    const [searchTerm, setSearchTerm] = useState('');
-    const [selectedCertificate, setSelectedCertificate] = useState(null);
-    const [filter, setFilter] = useState('all');
-
-    const filteredCertificates = certificatesData.filter(cert => {
-        const matchesSearch =
-            cert.title.toLowerCase().includes(searchTerm.toLowerCase()) ||
-            cert.institution.toLowerCase().includes(searchTerm.toLowerCase()) ||
-            (cert.description &&
-                cert.description
-                    .toLowerCase()
-                    .includes(searchTerm.toLowerCase()));
-
-        const matchesFilter = filter === 'all' || cert.type === filter;
-
-        return matchesSearch && matchesFilter;
-    });
-
-    const categories = ['all', ...new Set(certificatesData.map(c => c.type))];
-
-    return (
-        <section className="min-h-screen bg-black pt-40 pb-20 px-4 sm:px-6 lg:px-8 font-sans">
-            <div className="max-w-7xl mx-auto">
-                <motion.div
-                    initial={{ opacity: 0, y: -20 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    className="text-center mb-12"
-                >
-                    <h1 className="text-4xl font-light text-white mb-4 tracking-tight">
-                        Academic Achievements
-                    </h1>
-                    <p className="text-gray-400 font-light max-w-2xl mx-auto mb-8">
-                        My qualifications, certifications, and extracurricular
-                        activities.
-                    </p>
-
-                    <div className="relative max-w-md mx-auto mb-8">
-                        <input
-                            type="text"
-                            placeholder="Search achievements..."
-                            value={searchTerm}
-                            onChange={e => setSearchTerm(e.target.value)}
-                            className="w-full bg-transparent border border-gray-800 rounded-full py-3 pl-12 pr-4 text-white placeholder-gray-500 focus:outline-none focus:border-white focus:ring-0 transition-colors font-light"
-                        />
-                        <Search
-                            className="absolute left-4 top-1/2 transform -translate-y-1/2 text-gray-500"
-                            size={18}
-                        />
-                    </div>
-
-                    <div className="flex justify-center gap-3 flex-wrap">
-                        {categories.map(category => (
-                            <button
-                                key={category}
-                                onClick={() => setFilter(category)}
-                                className={`px-5 py-2 rounded-full text-sm font-light transition-all duration-300 border capitalize ${
-                                    filter === category
-                                        ? 'bg-white text-black border-white'
-                                        : 'bg-transparent text-gray-400 border-gray-800 hover:border-gray-600 hover:text-white'
-                                }`}
-                            >
-                                {category}
-                            </button>
-                        ))}
-                    </div>
-                </motion.div>
-
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-                    {filteredCertificates.map(cert => (
-                        <CertificateCard
-                            key={cert.id}
-                            certificate={cert}
-                            onClick={setSelectedCertificate}
-                        />
-                    ))}
-                </div>
-
-                {filteredCertificates.length === 0 && (
-                    <div className="text-center py-12">
-                        <p className="text-gray-500 font-light">
-                            No achievements found matching your criteria.
-                        </p>
-                    </div>
-                )}
-
-                {selectedCertificate && (
-                    <CertificateModal
-                        certificate={selectedCertificate}
-                        onClose={() => setSelectedCertificate(null)}
-                    />
-                )}
-            </div>
-        </section>
-    );
-};
-
-export default Certificates;
