+++ src/data/projects.ts
@@ -1,70 +1,15 @@
-import { useState } from 'react';
-import { motion } from 'framer-motion';
-import {
-    Github,
-    ExternalLink,
-    Code2,
-    Zap,
-    Palette,
-    Triangle,
-    Database,
-    Hexagon,
-    Youtube,
-    FileText,
-    FileType,
-    FileJson,
-    Brackets,
-    Braces,
-    SquareCode,
-    Brain,
-    Terminal,
-    Mic,
-    MapPin,
-    MessageCircle,
-    Train
-} from 'lucide-react';
-import agrolearn from '../assets/projects/AgroLearn.png';
-import quickreadme from '../assets/projects/quickreadme.png';
-import elementaryschoolwebsite from '../assets/projects/elementaryschoolwebsite.png';
-import tutortime from '../assets/projects/tutortime.png';
-import webdevclasses from '../assets/projects/webDevClasses.png';
-import historytestsite from '../assets/projects/historyTestStudyPortal.png';
-import speakScribeImage from '../assets/projects/speakScribe.png';
-import girlTalkAiImage from '../assets/projects/girlTalkAI.png';
-import encantosDoForno from '../assets/projects/encantosDoForno.png';
+import type { Project } from '@/types';
+import agrolearn from '@/assets/projects/AgroLearn.png';
+import quickreadme from '@/assets/projects/quickreadme.png';
+import elementaryschoolwebsite from '@/assets/projects/elementaryschoolwebsite.png';
+import tutortime from '@/assets/projects/tutortime.png';
+import webdevclasses from '@/assets/projects/webDevClasses.png';
+import historytestsite from '@/assets/projects/historyTestStudyPortal.png';
+import speakScribeImage from '@/assets/projects/speakScribe.png';
+import girlTalkAiImage from '@/assets/projects/girlTalkAI.png';
+import encantosDoForno from '@/assets/projects/encantosDoForno.png';
 
-const techIcons = {
-    React: <Code2 size={14} className="inline-block text-current" />,
-    Vite: <Zap size={14} className="inline-block text-current" />,
-    'Tailwind CSS': <Palette size={14} className="inline-block text-current" />,
-    Vercel: <Triangle size={14} className="inline-block text-current" />,
-    Python: <Terminal size={14} className="inline-block text-current" />,
-    HandleBars: <Brackets size={14} className="inline-block text-current" />,
-    'Node.js': <FileJson size={14} className="inline-block text-current" />,
-    'Express.js': (
-        <SquareCode size={14} className="inline-block text-current" />
-    ),
-    MySQL: <Database size={14} className="inline-block text-current" />,
-    Sequelize: <Database size={14} className="inline-block text-current" />,
-    HTML: <Code2 size={14} className="inline-block text-current" />,
-    CSS: <Palette size={14} className="inline-block text-current" />,
-    JavaScript: <Braces size={14} className="inline-block text-current" />,
-    TypeScript: <FileType size={14} className="inline-block text-current" />,
-    Bootstrap: <Palette size={14} className="inline-block text-current" />,
-    'Community Service': (
-        <Hexagon size={14} className="inline-block text-current" />
-    ),
-    Supabase: <Database size={14} className="inline-block text-current" />,
-    'GPT-3.5 Turbo': <Brain size={14} className="inline-block text-current" />,
-    Firestore: <Database size={14} className="inline-block text-current" />,
-    'Murf.ai': <Mic size={14} className="inline-block text-current" />,
-    Llama: <Brain size={14} className="inline-block text-current" />,
-    WhatsApp: <MessageCircle size={14} className="inline-block text-current" />,
-    Maps: <MapPin size={14} className="inline-block text-current" />,
-    Railway: <Train size={14} className="inline-block text-current" />
-};
-
-const projectsData = [
+export const projects: Project[] = [
     {
         id: 8,
         title: 'GirlTalk AI',
@@ -268,284 +213,3 @@ const projectsData = [
         ]
     }
 ];
-
-const ProjectCard = ({ project, onClick }) => (
-    <motion.div
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        whileHover={{ y: -5 }}
-        className="bg-dark-secondary/30 border border-gray-800 rounded-2xl overflow-hidden hover:border-gray-600 transition-all duration-300 group cursor-pointer"
-        onClick={() => onClick(project)}
-    >
-        <div className="relative aspect-video overflow-hidden">
-            <img
-                src={project.thumbnail}
-                alt={project.title}
-                className="w-full h-full object-cover transition-transform duration-500 group-hover:scale-105"
-            />
-            <div className="absolute inset-0 bg-black/60 opacity-0 group-hover:opacity-100 transition-opacity duration-300 flex items-center justify-center">
-                <span className="text-white font-light tracking-wide border border-white/30 px-4 py-2 rounded-full backdrop-blur-sm">
-                    View Details
-                </span>
-            </div>
-        </div>
-
-        <div className="p-6">
-            <h3 className="text-xl font-medium text-white mb-2">
-                {project.title}
-            </h3>
-            <p className="text-gray-400 text-sm font-light mb-4 line-clamp-2">
-                {project.shortDescription}
-            </p>
-
-            <div className="flex flex-wrap gap-2 mb-4">
-                {project.technologies.slice(0, 3).map(tech => (
-                    <span
-                        key={tech}
-                        className="px-2 py-1 bg-dark-primary border border-gray-800 rounded-lg text-xs text-gray-500 font-light flex items-center gap-1"
-                    >
-                        {techIcons[tech]} {tech}
-                    </span>
-                ))}
-                {project.technologies.length > 3 && (
-                    <span className="px-2 py-1 bg-dark-primary border border-gray-800 rounded-lg text-xs text-gray-500 font-light">
-                        +{project.technologies.length - 3}
-                    </span>
-                )}
-            </div>
-
-            <div className="flex gap-4 pt-2 border-t border-gray-800/50">
-                <a
-                    href={project.github}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    onClick={e => e.stopPropagation()}
-                    className="flex items-center gap-2 text-xs text-gray-500 hover:text-white transition-colors"
-                >
-                    <Github size={14} /> Code
-                </a>
-                {project.demo && (
-                    <a
-                        href={project.demo}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        onClick={e => e.stopPropagation()}
-                        className="flex items-center gap-2 text-xs text-gray-500 hover:text-white transition-colors"
-                    >
-                        <ExternalLink size={14} /> Demo
-                    </a>
-                )}
-            </div>
-        </div>
-    </motion.div>
-);
-
-const ProjectModal = ({ project, onClose }) => (
-    <div
-        className="fixed inset-0 bg-black/80 backdrop-blur-sm flex items-center justify-center z-50 p-4"
-        onClick={onClose}
-    >
-        <motion.div
-            initial={{ scale: 0.95, opacity: 0 }}
-            animate={{ scale: 1, opacity: 1 }}
-            className="bg-dark-primary border border-gray-800 rounded-2xl max-w-4xl w-full max-h-[90vh] overflow-y-auto shadow-2xl"
-            onClick={e => e.stopPropagation()}
-        >
-            <div className="relative h-64 sm:h-80">
-                <img
-                    src={project.image}
-                    alt={project.title}
-                    className="w-full h-full object-cover"
-                />
-                <div className="absolute inset-0 bg-gradient-to-t from-dark-primary to-transparent"></div>
-                <button
-                    onClick={onClose}
-                    className="absolute top-4 right-4 p-2 bg-black/50 rounded-full text-white hover:bg-black/70 transition-colors"
-                >
-                    <svg
-                        className="w-6 h-6"
-                        fill="none"
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                    >
-                        <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth="2"
-                            d="M6 18L18 6M6 6l12 12"
-                        ></path>
-                    </svg>
-                </button>
-            </div>
-
-            <div className="p-8 -mt-12 relative">
-                <h2 className="text-3xl font-light text-white mb-4">
-                    {project.title}
-                </h2>
-
-                <div className="flex flex-wrap gap-2 mb-6">
-                    {project.technologies.map(tech => (
-                        <span
-                            key={tech}
-                            className="px-3 py-1 bg-dark-secondary/50 border border-gray-700 rounded-full text-sm text-gray-300 font-light flex items-center gap-2"
-                        >
-                            {techIcons[tech]} {tech}
-                        </span>
-                    ))}
-                </div>
-
-                <div className="grid md:grid-cols-3 gap-8">
-                    <div className="md:col-span-2 space-y-6">
-                        <div>
-                            <h3 className="text-sm font-medium text-gray-500 uppercase tracking-wider mb-3">
-                                Overview
-                            </h3>
-                            <p className="text-gray-300 font-light leading-relaxed">
-                                {project.fullDescription}
-                            </p>
-                        </div>
-
-                        <div>
-                            <h3 className="text-sm font-medium text-gray-500 uppercase tracking-wider mb-3">
-                                Key Highlights
-                            </h3>
-                            <ul className="space-y-2">
-                                {project.highlights.map((highlight, index) => (
-                                    <li
-                                        key={index}
-                                        className="flex items-start gap-3 text-gray-400 font-light"
-                                    >
-                                        <span className="mt-1.5 w-1.5 h-1.5 rounded-full bg-gray-600 flex-shrink-0"></span>
-                                        <span>{highlight}</span>
-                                    </li>
-                                ))}
-                            </ul>
-                        </div>
-                    </div>
-
-                    <div className="space-y-4">
-                        <h3 className="text-sm font-medium text-gray-500 uppercase tracking-wider mb-3">
-                            Links
-                        </h3>
-                        <a
-                            href={project.github}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            className="flex items-center justify-center gap-2 w-full py-3 rounded-xl bg-white text-black font-medium hover:bg-gray-200 transition-colors"
-                        >
-                            <Github size={18} /> View Code
-                        </a>
-                        {project.demo && (
-                            <a
-                                href={project.demo}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="flex items-center justify-center gap-2 w-full py-3 rounded-xl border border-gray-700 text-white hover:bg-gray-800 transition-colors"
-                            >
-                                <ExternalLink size={18} /> Live Demo
-                            </a>
-                        )}
-                        {project.devpost && (
-                            <a
-                                href={project.devpost}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="flex items-center justify-center gap-2 w-full py-3 rounded-xl border border-gray-700 text-white hover:bg-gray-800 transition-colors"
-                            >
-                                <Hexagon size={18} /> Devpost
-                            </a>
-                        )}
-                        {project.youtube && (
-                            <a
-                                href={project.youtube}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="flex items-center justify-center gap-2 w-full py-3 rounded-xl border border-gray-700 text-white hover:bg-gray-800 transition-colors"
-                            >
-                                <Youtube size={18} /> Video
-                            </a>
-                        )}
-                        {project.slides && (
-                            <a
-                                href={project.slides}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="flex items-center justify-center gap-2 w-full py-3 rounded-xl border border-gray-700 text-white hover:bg-gray-800 transition-colors"
-                            >
-                                <FileText size={18} /> Slides
-                            </a>
-                        )}
-                    </div>
-                </div>
-            </div>
-        </motion.div>
-    </div>
-);
-
-const Projects = () => {
-    const [selectedProject, setSelectedProject] = useState(null);
-    const [filter, setFilter] = useState('all');
-
-    const categories = ['all', ...new Set(projectsData.map(p => p.category))];
-
-    const filteredProjects =
-        filter === 'all'
-            ? projectsData
-            : projectsData.filter(p => p.category === filter);
-
-    return (
-        <section className="min-h-screen bg-black pt-40 pb-20 px-4 sm:px-6 lg:px-8 font-sans">
-            <div className="max-w-7xl mx-auto">
-                <motion.div
-                    initial={{ opacity: 0, y: -20 }}
-                    animate={{ opacity: 1, y: 0 }}
-                    className="text-center mb-16"
-                >
-                    <h1 className="text-4xl font-light text-white mb-4 tracking-tight">
-                        Selected Projects
-                    </h1>
-                    <p className="text-gray-400 font-light max-w-2xl mx-auto">
-                        A collection of my work in web development, AI, and open
-                        source contributions.
-                    </p>
-                </motion.div>
-
-                <div className="flex justify-center gap-3 mb-12 flex-wrap">
-                    {categories.map(category => (
-                        <button
-                            key={category}
-                            onClick={() => setFilter(category)}
-                            className={`px-5 py-2 rounded-full text-sm font-light transition-all duration-300 border ${
-                                filter === category
-                                    ? 'bg-white text-black border-white'
-                                    : 'bg-transparent text-gray-400 border-gray-800 hover:border-gray-600 hover:text-white'
-                            }`}
-                        >
-                            {category.charAt(0).toUpperCase() +
-                                category.slice(1)}
-                        </button>
-                    ))}
-                </div>
-
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
-                    {filteredProjects.map(project => (
-                        <ProjectCard
-                            key={project.id}
-                            project={project}
-                            onClick={setSelectedProject}
-                        />
-                    ))}
-                </div>
-
-                {selectedProject && (
-                    <ProjectModal
-                        project={selectedProject}
-                        onClose={() => setSelectedProject(null)}
-                    />
-                )}
-            </div>
-        </section>
-    );
-};
-
-export default Projects;
