+++ src/data/timeline.ts
@@ -1,0 +1,29 @@
+import type { TimelineEntry } from '@/types';
+
+export const timeline: TimelineEntry[] = [
+    {
+        year: '2016',
+        description:
+            'Started programming at the age of nine, creating small projects to solve problems and learn new skills.'
+    },
+    {
+        year: '2022',
+        description:
+            'Accepted into the most competitive and prestigious free high school program in Brazil, Federal Institute of Science and Technology, excelling in various subjects.'
+    },
+    {
+        year: '2024',
+        description:
+            'Developed TutorTime, my first full-stack application and Capstone Project for my Informatics Technician course (completed ahead of schedule in my junior year). Achieved a dream by participating in my first hackathon at an Ivy League, PennApps at UPenn, marking my first academic trip abroad. I successfully organized a fundraiser campaign to raise the necessary funds for this independent journey.'
+    },
+    {
+        year: '2024-2025',
+        description:
+            'Completed an intensive programming challenge by coding every day from September 22, 2024 to January 16, 2025. This consistent practice strengthened my skills, expanded my portfolio, and demonstrated my dedication to continuous improvement in software development.'
+    },
+    {
+        year: '2025',
+        description:
+            'Deepened my skills on frontend and backend. Attempted to the Hackathon that changes my life - UCBerkey AI Hackathon, and since then, developing my skills on Applied AI, AI Agents, etc...'
+    }
+];
