--- src/main.jsx
@@ -1,30 +1,0 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
-import './index.css';
-import App from './App.jsx';
-
-// Get base path from Vite configuration (injected at build time)
-const basename = import.meta.env.BASE_URL;
-
-// Register Service Worker
-if ('serviceWorker' in navigator) {
-    window.addEventListener('load', () => {
-        navigator.serviceWorker
-            .register(`${basename}sw.js`)
-            .then(() => {
-                // Service Worker registered successfully
-            })
-            .catch(() => {
-                // Service Worker registration failed
-            });
-    });
-}
-
-createRoot(document.getElementById('root')).render(
-    <StrictMode>
-        <BrowserRouter basename={basename}>
-            <App />
-        </BrowserRouter>
-    </StrictMode>
-);
