+++ src/main.tsx
@@ -1,0 +1,19 @@
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import './index.css';
+import App from './App.tsx';
+
+// Base path injected at build time (/ on Vercel, /luis.hsa on GitHub Pages).
+const basename = import.meta.env.BASE_URL;
+
+const rootElement = document.getElementById('root');
+if (!rootElement) throw new Error('Root element #root not found');
+
+createRoot(rootElement).render(
+    <StrictMode>
+        <BrowserRouter basename={basename}>
+            <App />
+        </BrowserRouter>
+    </StrictMode>
+);
