+++ vite.config.ts
@@ -1,13 +1,19 @@
 import { defineConfig, loadEnv } from 'vite';
 import react from '@vitejs/plugin-react';
+import { fileURLToPath, URL } from 'node:url';
 
 // https://vite.dev/config/
 export default defineConfig(({ mode }) => {
     // Load environment variables based on current mode
-    const env = loadEnv(mode, import.meta.dirname || '.', '');
+    const env = loadEnv(mode, process.cwd(), '');
 
     return {
         plugins: [react()],
+        resolve: {
+            alias: {
+                '@': fileURLToPath(new URL('./src', import.meta.url))
+            }
+        },
         // Use '/' for Vercel (root deployment), '/luis.hsa' for GitHub Pages (subfolder)
         base: env.VITE_BASE_PATH || '/',
         server: {
